# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Phonebook API
  description: Simple internship phonebook project.
  version: 0.0.1

servers:
  - url: http://localhost:8080
    description: Local server.

tags:
  - name: record
    description: Phonebook record.

paths:
  /record:
    get:
      tags:
        - record
      summary: Returns a list of records.
      parameters:
        - in: query
          name: sortBy
          schema:
            type: string
          description: Field to sort by.
        - in: query
          name: orderBy
          schema:
            type: string
          description: Direction to order by.
      responses:
        '200':
          description: A JSON array of phonebook records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Record'
  /record/add:
    post:
      tags:
        - record
      summary: Adds record to phonebook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Record'
      responses:
        '200':
          description: Successfully added.
          content:
            application/json:
              schema:
                type: string

  /record/delete:
    delete:
      tags:
        - record
      summary: Removes record to phonebook.
      parameters:
        - in: query
          name: phoneNumber
          schema:
            type: string
          description: Phone number to remove from phonebook.
      responses:
        '200':
          description: Response of successfully added or no such phone number.
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    Record:
      type: object
      required:
        - id
        - firstname
        - lastname
        - phonenumber
      properties:
        id:
          type: integer
          format: int64
          example: 1
        firstname:
          type: string
          example: Andrew
        lastname:
          type: string
          example: Nowak
        phonenumber:
          type: string
          example: 123456789
      xml:
        name: record
  requestBodies:
    Record:
      description: Record to add.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Record'